Massive refactoring of eyes.py Still work in progress
Tried to split logic into different classes.
Eye animation went to EyesModel + lots of supporting classes.
Rendering is performed by Renderer which is quite dumb now.
Copying image to OLED screens is now performed by separate threads
(one per OLED) so rendering of the next frame can be done in parallel.

It is still a work in progress and I mostly broke support for
controlling eyes with ADC and buttons but I hope it is temporary.
At least I kept in mind all these things when I was crafting EyesModel.